Display of all plots;

+++ 29_Financial/Charting_09_ScottPlot_MultipleCharts_AutoProps/Sample/DataResultModel.cs
@@ -7,7 +7,7 @@ namespace Sample
         public Guid DataGuid { get; set; }
         public string IndexName { get; set; }
         public string DataType { get; set; }
-        public decimal ResultValue { get; set; }
+        public double ResultValue { get; set; }
         public DateTime UtcTimeStamp { get; set; }
         public bool Release { get; set; }
     }
